support

+++ app/controllers/CpanelEventController.php
@@ -124,7 +124,8 @@
         //Getting and uploading logo file.
         if(Input::hasFile('image_logo'))
         {
-            unlink('../..'.$vent->image_logo);
+          if(file_exists('../..'. $vent->image_logo))
+              unlink('../..'.$vent->image_logo);
             $file = Input::file('image_logo');
             $img_path = '../../images/events';
             $file_name = $vent->name . '-' . time() . '.' . $file->getClientOriginalExtension();
@@ -154,7 +155,7 @@
         {
             $vent->Sponsers()->detach($sponserofvent->id);
         }
-
+     if(file_exists('../..'. $vent->image_logo))
         unlink('../..'. $vent->image_logo);
 
         //Deleting.
